css/css-gcpm/inheritance.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Inheritance of CSS Generated Content for Paged Media properties</title>
<link rel="help" href="https://drafts.csswg.org/css-gcpm/#property-index">
<meta name="assert" content="Properties do not inherit.">
<meta name="assert" content="Properties have initial values according to the spec.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/inheritance-testcommon.js"></script>
</head>
<body>
<div id="reference"></div>
<div id="container">
  <div id="target"></div>
</div>
<script>
assert_not_inherited('footnote-display', 'block', 'inline');
assert_not_inherited('footnote-policy', 'auto', 'block');
assert_not_inherited('running', 'none', 'RuNnInG');
assert_not_inherited('string-set', 'none', 'StRiNg counter(SeT)');
</script>
</body>
</html>
